Fix Windows Stage-2 compilation
After the multi-stage refactor, Windows builds fails because the Go
Rebuild Urself™ Technology has a bug for `cl.exe`.

It was relying on `cl.exe` behaviour of defaulting the output file to
the source file with `exe` extension.

This works fine in the classical usage of the Go Rebuild Urself™
Technology (i.e. rename old file, then build a new one).

But the new shiny usage (multi-stage builder) won't work because we
need different output names. Most cases, we use the `binary_path`
parameter of the `NOB_REBUILD_URSELF` macro, but in the nichiest of
compilers was trying to build a new `nob.exe` instead - which fails
because Windows has `nob.exe` locked while it executes.